stratified_resample shifted
